+++ docs/CHANGELOG.md
@@ -6,7 +6,10 @@ Released: ???
 
 Features:
 - Added Sea creatures count threshold - GALATEA setting (for MC 1.21).
-- Removed party alerts on Rare & Epic Flying Fish.
+- Removed party alerts on Rare & Epic Flying Fish pet drop.
+- Removed party alert on Epic Baby Yeti pet drop.
+- Removed count of dropped Baby Yeti pets from Jerry Workshop tracker.
+  - I had to remove it because now is hard to track dropped pets - Hypixel removed RARE DROP message for some pets.
 
 Bugfixes:
 -